* rename src/api to src/contracts (1/2)
