+++ src/contracts/FarmStream.Announce.ts
