+++ src/contracts/Staker.API.ts
