+++ src/contracts/Staker.Fetch.ts
