+++ src/contracts/Staker.Stake.ts
